Move forgot password handlers to separate pkg
