+++ pkg/auth/handler/forgotpwdhandler/forgot_password.go
@@ -1,4 +1,4 @@
-package handler
+package forgotpwdhandler
 
 import (
 	"encoding/json"
