+++ pkg/auth/handler/forgotpwdhandler/forgot_password_test.go
@@ -1,4 +1,4 @@
-package handler
+package forgotpwdhandler
 
 import (
 	"net/http"
