+++ pkg/auth/handler/forgotpwdhandler/handler.go
@@ -1,0 +1,7 @@
+package forgotpwdhandler
+
+import "time"
+
+var (
+	timeNow = func() time.Time { return time.Now().UTC() }
+)
