+++ pkg/auth/handler/forgotpwdhandler/reset.go
@@ -1,4 +1,4 @@
-package handler
+package forgotpwdhandler
 
 import (
 	"encoding/json"
