+++ pkg/auth/handler/forgotpwdhandler/reset_test.go
@@ -1,4 +1,4 @@
-package handler
+package forgotpwdhandler
 
 import (
 	"encoding/json"
